Add files via upload
Postal Savings Bank of China Co. Ltd. also known as PSBC is a commercial retail bank founded in 2007 and headquartered in Beijing that deals with Banking and Insurance. The company is intending to expand its service across the Shaanxi province having the first branch to be located in Xian, ankang . The company has secured a four- floor building to operate within ankang city. Therefore, the company would like to allow sourcing the knowledge from a group of final year student form ankang university to design and implement their company network. So, the requirements, model and the design and implement the network based on the company’s needs.